Attempt to fix Macos build

+++ picows/picows_compat.h
@@ -120,6 +120,6 @@
     {
         struct timespec ts;
         clock_gettime(CLOCK_MONOTONIC, &ts);
-        return ts.tv_sec + ts.tv_nsec / 1e9;
+        return (double)ts.tv_sec + (double)ts.tv_nsec / 1e9;
     }
 #endif
